added OpenMP update, removed SLIError from sli_neuron, added missing static_cast to ppd_sup_generator, added comment on using in stdp_triplet_connection
git-svn-id: https://www.nest-initiative.uni-freiburg.de/nest/branches/10kproject@9262 4fa6ec7b-3718-0410-b4d6-90ca9082ef84

+++ developer/ppd_sup_generator.cpp
@@ -47,7 +47,7 @@ nest::ppd_sup_generator::FastBinomialRandomDev_::FastBinomialRandomDev_(size_t n
         f_[i+1] = 0.0;
         j = 1;
         while (j<=i){
-            f_[i+1] += std::log(j);
+	  f_[i+1] += std::log(static_cast<double>(j));
             j++;
             }
         i++;
